backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework-simplejwt==5.3.1
Pillow==10.2.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
django-storages==1.14.2
boto3==1.34.34 